+++ resolwe_docker_images/rnaseq/packages-manual/cufflinks.sh
@@ -7,7 +7,7 @@ download_and_verify \
     cufflinks \
     2.2.1 \
     9615bba771e6c89546644c653a99127e49c4453396f2cc7f98b6363c157bb961 \
-    https://drive.google.com/uc?id=14jDS0XP4ZOd2uXf735YWUSMZyLarUIvV#.elf \
+    https://raw.githubusercontent.com/genialis/resolwe-docker-images/43374488699a9d9dd6f158152bf9499b141824fe/resolwe_docker_images/rnaseq/downloaded_assets/cufflinks-2.2.1.tar.bz2 \
     cufflinks \
     cufflinks-2.2.1.tar.bz2