+++ resolwe_docker_images/rnaseq/packages-manual/ercc.sh
@@ -10,6 +10,6 @@ download_and_verify \
     assets \
     ERCC_table \
     5151b394332477301851404d37b907cb6f27c9aaa2569e288531a839d6ca2b3b \
-    https://drive.google.com/uc?id=193PpxSoNhFwYUpTMh9nGt44K72tc0F9f \
+    https://raw.githubusercontent.com/genialis/resolwe-docker-images/43374488699a9d9dd6f158152bf9499b141824fe/resolwe_docker_images/rnaseq/downloaded_assets/ERCC_table.txt \
     '' \
     ERCC_table.txt